fact(bgp_policies_test): move BGP policy tests into their own file
